requirements.txt
requests
beautifulsoup4
dnspython
Pillow
pyspellchecker
flask
playwright
